Add board to the game - it's size is based on the value that user provides

+++ features/queengame/src/main/java/com/jaxjack/queens/features/queengame/configuration/navigation/GameConfigurationEntry.kt
@@ -5,14 +5,8 @@ import androidx.navigation3.runtime.EntryProviderScope
 import androidx.navigation3.runtime.NavKey
 import com.jaxjack.queens.features.queengame.configuration.GameConfigurationScreen
 
-/**
- * Registers the board setup destination.
- *
- * The caller decides what navigating away means; the screen itself stays
- * internal to this module.
- */
 fun EntryProviderScope<NavKey>.gameConfigurationEntry(
-    onPlayClick: () -> Unit,
+    onPlayClick: (Int) -> Unit,
 ) {
     entry<GameConfigurationKey> {
         GameConfigurationScreen(
